pyproject.toml
[tool.poetry]
name = "django-blog"
version = "0.1.0"
description = ""
authors = ["Timur Safin <timur@getgain.ru>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2.9"
Pillow = "^7.0.0"
django-filter = "^2.2.0"

[tool.poetry.dev-dependencies]
wemake-python-styleguide = "^0.13.4"
pytest-django = "^3.8.0"
mypy = "^0.761"
autopep8 = "^1.5"
mixer = "^6.1.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
